requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMongo==4.6.0
bcrypt==4.1.2
PyJWT==2.8.0
flask-bcrypt==1.0.1
Flask-JWT-Extended==4.6.0

# Security and Rate Limiting
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
Werkzeug==3.0.1

# Input Validation and Sanitization
marshmallow==3.20.2
flask-marshmallow==0.15.0
bleach==6.1.0
validators==0.22.0

# Logging and Monitoring
structlog==23.2.0

# Additional Security Headers
Flask-SeaSurf==1.1.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1